third prop row to separate component - dnd wrap
